Переносит функции между модулями
