+++ js/data.js
@@ -1,11 +1,14 @@
 const GET_DATA_URL = 'https://22.javascript.pages.academy/keksobooking/data';
 const SEND_DATA_URL = 'https://22.javascript.pages.academy/keksobooking';
 
-const getData = (onSuccess) => {
+const getData = (onSuccess, onFail) => {
   fetch(GET_DATA_URL)
     .then((response) => response.json())
     .then((offers) => {
       onSuccess(offers);
+    })
+    .catch(() => {
+      onFail('При загрузке данных с сервера произошла ошибка');
     });
 };
