+++ js/form.js
@@ -1,6 +1,6 @@
 import { TYPES } from './card.js';
 import { sendData } from './data.js';
-import { isEscEvent } from './util.js';
+import { showSuccessMessage, showErrorMessage } from './util.js';
 
 const roomsCapacity = {
   1: ['1'],
@@ -61,41 +61,6 @@ const validateRoomSelect = () => {
 validateRoomSelect();
 
 roomSelect.addEventListener('change', validateRoomSelect);
-
-const closePopup = () => {
-  if (document.querySelector('.success')) {
-    document.querySelector('.success').remove();
-  }
-  if (document.querySelector('.error')) {
-    document.querySelector('.error').remove();
-  }
-  document.removeEventListener('keydown', onPopupEscKeydown);
-}
-
-const onPopupEscKeydown = (evt) => {
-  if (isEscEvent(evt)) {
-    evt.preventDefault();
-    closePopup();
-  }
-}
-
-const showSuccessMessage = () => {
-  const successTemplate = document.querySelector('#success').content.querySelector('.success');
-  const successMessage = successTemplate.cloneNode(true);
-  successMessage.style.zIndex = 1000;
-  document.querySelector('main').append(successMessage);
-  document.addEventListener('keydown', onPopupEscKeydown);
-  document.addEventListener('click', closePopup);
-};
-
-const showErrorMessage = () => {
-  const errorTemplate = document.querySelector('#error').content.querySelector('.error');
-  const errorMessage = errorTemplate.cloneNode(true);
-  errorMessage.style.zIndex = 1000;
-  document.querySelector('main').append(errorMessage);
-  document.addEventListener('keydown', onPopupEscKeydown);
-  document.addEventListener('click', closePopup);
-};
 
 const onSubmitAdForm = (onSuccess, onFail) => {
   adForm.addEventListener('submit', (evt) => {
