+++ js/util.js
@@ -66,6 +66,44 @@ const isEscEvent = (evt) => {
   return evt.key === 'Escape' || evt.key === 'Esc';
 };
 
+const closePopup = () => {
+  if (document.querySelector('.success')) {
+    document.querySelector('.success').remove();
+  }
+  if (document.querySelector('.error')) {
+    document.querySelector('.error').remove();
+  }
+  document.removeEventListener('keydown', onPopupEscKeydown);
+}
+
+const onPopupEscKeydown = (evt) => {
+  if (isEscEvent(evt)) {
+    evt.preventDefault();
+    closePopup();
+  }
+}
+
+const showSuccessMessage = () => {
+  const successTemplate = document.querySelector('#success').content.querySelector('.success');
+  const successMessage = successTemplate.cloneNode(true);
+  successMessage.style.zIndex = 1000;
+  document.querySelector('main').append(successMessage);
+  document.addEventListener('keydown', onPopupEscKeydown);
+  document.addEventListener('click', closePopup);
+};
+
+const showErrorMessage = (message) => {
+  const errorTemplate = document.querySelector('#error').content.querySelector('.error');
+  const errorMessage = errorTemplate.cloneNode(true);
+  errorMessage.style.zIndex = 1000;
+  if(message) {
+    errorMessage.querySelector('p').textContent = message;
+  }
+  document.querySelector('main').append(errorMessage);
+  document.addEventListener('keydown', onPopupEscKeydown);
+  document.addEventListener('click', closePopup);
+};
+
 export {
   getRandomNumber,
   getRandomArrayElement,
@@ -74,5 +112,6 @@ export {
   createOfferPhotos,
   createOfferFeatures,
   getAddress,
-  isEscEvent
+  showSuccessMessage,
+  showErrorMessage
 }
